+++ src/app/snippets/mod.rs
@@ -1,4 +1,6 @@
 mod snippets;
 
-pub(crate) use snippets::Snippet;
 pub(crate) use snippets::standard_postgres_snippets;
+pub(crate) use snippets::Snippet;
+pub(crate) use snippets::SnippetDatabase;
+pub(crate) use snippets::SnippetsConfig;